Agregar columna 'dark_mode' a la tabla 'users' y eliminar migración de 'notifications'.

+++ database/migrations/2025_03_26_113805_add_dark_mode_to_users_table.php
@@ -14,8 +14,10 @@ return new class extends Migration {
 
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('dark_mode');
-        });
+        if (Schema::hasColumn('users', 'dark_mode')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('dark_mode');
+            });
+        }
     }
 };
